extract widgets to reduce lines in wallet_home
