+++ lib/screens/wallet_home.dart
@@ -11,8 +11,9 @@ import 'package:peercoin/tools/auth.dart';
 import 'package:peercoin/widgets/app_drawer.dart';
 import 'package:peercoin/widgets/loading_indicator.dart';
 import 'package:peercoin/widgets/wallet_content_switch.dart';
+import 'package:peercoin/widgets/wallet_home_connection.dart';
+import 'package:peercoin/widgets/wallet_home_qr.dart';
 import 'package:provider/provider.dart';
-import 'package:qr_flutter/qr_flutter.dart';
 
 class WalletHomeScreen extends StatefulWidget {
   @override
@@ -292,78 +293,10 @@ class _WalletHomeState extends State<WalletHomeScreen>
                             : Container(),
                       ],
                     ),
-                    _unusedAddress == ""
-                        ? SizedBox(height: 60, width: 60)
-                        : InkWell(
-                            onTap: () {
-                              showDialog(
-                                  context: context,
-                                  builder: (BuildContext context) {
-                                    return SimpleDialog(children: [
-                                      Center(
-                                          child: Column(children: [
-                                        SizedBox(
-                                          height: MediaQuery.of(context)
-                                                  .size
-                                                  .height *
-                                              0.33,
-                                          width: MediaQuery.of(context)
-                                                  .size
-                                                  .width *
-                                              1,
-                                          child: Center(
-                                            child: QrImage(
-                                              data: _unusedAddress,
-                                              backgroundColor: Colors.white,
-                                              foregroundColor: Colors.black,
-                                            ),
-                                          ),
-                                        ),
-                                        Padding(
-                                          padding: const EdgeInsets.all(8.0),
-                                          child: FittedBox(
-                                            child: SelectableText(
-                                              _unusedAddress,
-                                              style: TextStyle(
-                                                  fontWeight: FontWeight.bold),
-                                            ),
-                                          ),
-                                        )
-                                      ]))
-                                    ]);
-                                  });
-                            },
-                            child: QrImage(
-                              data: _unusedAddress,
-                              size: 60.0,
-                              padding: EdgeInsets.all(1),
-                              backgroundColor: Colors.white,
-                              foregroundColor: Colors.black,
-                            ),
-                          )
+                    WalletHomeQr(_unusedAddress)
                   ],
                 ),
-                _connectionState == "connected"
-                    ? Row(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: [
-                          Icon(
-                            Icons.sync,
-                            color: Theme.of(context).primaryColor,
-                          ),
-                          Text(
-                            AppLocalizations.instance
-                                .translate('wallet_connected'),
-                            style: TextStyle(
-                                color: Theme.of(context).accentColor,
-                                fontSize: 12),
-                          ),
-                        ],
-                      )
-                    : Padding(
-                        padding: const EdgeInsets.all(10.0),
-                        child: LoadingIndicator(),
-                      ),
+                WalletHomeConnection(_connectionState),
                 Divider(),
                 WalletContentSwitch(
                     pageIndex: _pageIndex,
